Deprecate the integration to display a notification when Premium is not up-to-date
